float not double

+++ layers/pe_layer.py
@@ -261,7 +261,7 @@ class PELayer(nn.Module):
         if self.gape_normalize_mat:
             A = g.adjacency_matrix_scipy(return_edge_ids=False).astype(float)
             D = sp.sparse.diags(dgl.backend.asnumpy(g.in_degrees()).clip(1) ** -1.0, dtype=float)
-            mat = torch.from_numpy((A * D).todense()).to(self.device)
+            mat = torch.from_numpy((A * D).todense()).to(self.device).type(torch.float)
 
         vec_init = self.stack_strategy(g.number_of_nodes()).to(self.device)
         # transition = torch.diag(self.pos_transitions[0])
